start avatar animation refactor
