+++ avatars/AnimationMapping.js
@@ -1,0 +1,12 @@
+export default class AnimationMapping {
+  constructor(animationTrackName, boneName, isTop, isPosition) {
+    this.animationTrackName = animationTrackName;
+    this.boneName = boneName;
+    this.isTop = isTop;
+    this.isPosition = isPosition;
+  }
+
+  clone() {
+    return new AnimationMapping(this.animationTrackName, this.boneName, this.isTop, this.isPosition);
+  }
+}
