+++ avatars/Emoter.js
@@ -1,0 +1,11 @@
+/* eslint-disable no-useless-constructor */
+
+export default class Emoter {
+  constructor() {
+
+  }
+
+  update() {
+
+  }
+}
