+++ avatars/Nodder.js
@@ -1,0 +1,10 @@
+export default class Nodder {
+  // eslint-disable-next-line no-useless-constructor
+  constructor() {
+
+  }
+
+  update() {
+
+  }
+}
