+++ avatars/avatars.js
@@ -1,5 +1,5 @@
 import * as THREE from 'three';
-import {VRMSpringBoneImporter, VRMLookAtApplyer, VRMCurveMapper} from '@pixiv/three-vrm/lib/three-vrm.module.js';
+import {VRMSpringBoneImporter} from '@pixiv/three-vrm/lib/three-vrm.module.js';
 import {fixSkeletonZForward} from './vrarmik/SkeletonUtils.js';
 import PoseManager from './vrarmik/PoseManager.js';
 import ShoulderTransforms from './vrarmik/ShoulderTransforms.js';
@@ -8,12 +8,10 @@ import {scene, camera} from '../renderer.js';
 import MicrophoneWorker from './microphone-worker.js';
 import {AudioRecognizer} from '../audio-recognizer.js';
 import {
-  angleDifference,
   // getVelocityDampingFactor,
   getNextPhysicsId,
 } from '../util.js';
 import easing from '../easing.js';
-import {zbdecode} from 'zjs/encoding.mjs';
 import Simplex from '../simplex-noise.js';
 import {
   crouchMaxTime,
@@ -41,29 +39,19 @@ import {
   getTailBones,
   getModelBones,
   // cloneModelBones,
-  decorateAnimation,
   // retargetAnimation,
   // animationBoneToModelBone,
 } from './util.mjs';
 import metaversefile from 'metaversefile';
 
-const localVector = new THREE.Vector3();
-const localVector2 = new THREE.Vector3();
-const localVector3 = new THREE.Vector3();
-const localVector4 = new THREE.Vector3();
-const localVector5 = new THREE.Vector3();
-// const localVector6 = new THREE.Vector3();
-const localQuaternion = new THREE.Quaternion();
-const localQuaternion2 = new THREE.Quaternion();
-const localQuaternion3 = new THREE.Quaternion();
-const localQuaternion4 = new THREE.Quaternion();
-const localQuaternion5 = new THREE.Quaternion();
-const localQuaternion6 = new THREE.Quaternion();
-const localEuler = new THREE.Euler(0, 0, 0, 'YXZ');
-const localEuler2 = new THREE.Euler(0, 0, 0, 'YXZ');
-const localMatrix = new THREE.Matrix4();
-const localMatrix2 = new THREE.Matrix4();
-const localPlane = new THREE.Plane();
+import { getFirstPersonCurves, getClosest2AnimationAngles, loadPromise, _findArmature, _getLerpFn, _applyAnimation} from './animationHelpers.js'
+
+import {animationMappingConfig} from './animationMappingConfig';
+
+import Emoter from './Emoter.js'
+import Blinker from './Blinker.js'
+import Nodder from './Nodder.js'
+import Looker from './Looker.js'
 
 const textEncoder = new TextEncoder();
 
@@ -99,39 +87,7 @@ const maxEyeTargetTime = 2000;
   };
 })(VRMSpringBoneImporter.prototype._createSpringBone); */
 
-function getFirstPersonCurves(vrmExtension) {
-  if (vrmExtension) {
-    const {firstPerson} = vrmExtension;
-    const {
-      lookAtHorizontalInner,
-      lookAtHorizontalOuter,
-      lookAtVerticalDown,
-      lookAtVerticalUp,
-      // lookAtTypeName,
-    } = firstPerson;
 
-    const DEG2RAD = Math.PI / 180; // THREE.MathUtils.DEG2RAD;
-    function _importCurveMapperBone(map) {
-      return new VRMCurveMapper(
-        typeof map.xRange === 'number' ? DEG2RAD * map.xRange : undefined,
-        typeof map.yRange === 'number' ? DEG2RAD * map.yRange : undefined,
-        map.curve,
-      );
-    }
-    const lookAtHorizontalInnerCurve = _importCurveMapperBone(lookAtHorizontalInner);
-    const lookAtHorizontalOuterCurve = _importCurveMapperBone(lookAtHorizontalOuter);
-    const lookAtVerticalDownCurve = _importCurveMapperBone(lookAtVerticalDown);
-    const lookAtVerticalUpCurve = _importCurveMapperBone(lookAtVerticalUp);
-    return {
-      lookAtHorizontalInnerCurve,
-      lookAtHorizontalOuterCurve,
-      lookAtVerticalDownCurve,
-      lookAtVerticalUpCurve,
-    };
-  } else {
-    return null;
-  }
-}
 
 const _makeSimplexes = numSimplexes => {
   const result = Array(numSimplexes);
@@ -142,16 +98,7 @@ const _makeSimplexes = numSimplexes => {
 };
 const simplexes = _makeSimplexes(5);
 
-const getClosest2AnimationAngles = (key, angle) => {
-  const animationAngleArray = animationsAngleArrays[key];
-  animationAngleArray.sort((a, b) => {
-    const aDistance = Math.abs(angleDifference(angle, a.angle));
-    const bDistance = Math.abs(angleDifference(angle, b.angle));
-    return aDistance - bDistance;
-  });
-  const closest2AnimationAngles = animationAngleArray.slice(0, 2);
-  return closest2AnimationAngles;
-};
+
 
 const upVector = new THREE.Vector3(0, 1, 0);
 const upRotation = new THREE.Quaternion().setFromAxisAngle(new THREE.Vector3(1, 0, 0), Math.PI*0.5);
@@ -171,388 +118,11 @@ const defaultNarutoRunAnimation = 'narutoRun';
 // const defaultHurtAnimation = 'pain_back';
 
 const infinityUpVector = new THREE.Vector3(0, Infinity, 0);
-// const crouchMagnitude = 0.2;
-/* const animationsSelectMap = {
-  crouch: {
-    'Crouch Idle.fbx': new THREE.Vector3(0, 0, 0),
-    'Sneaking Forward.fbx': new THREE.Vector3(0, 0, -crouchMagnitude),
-    'Sneaking Forward reverse.fbx': new THREE.Vector3(0, 0, crouchMagnitude),
-    'Crouched Sneaking Left.fbx': new THREE.Vector3(-crouchMagnitude, 0, 0),
-    'Crouched Sneaking Right.fbx': new THREE.Vector3(crouchMagnitude, 0, 0),
-  },
-  stand: {
-    'idle.fbx': new THREE.Vector3(0, 0, 0),
-    'jump.fbx': new THREE.Vector3(0, 1, 0),
+import {
+  animations,
+  animationStepIndices
+} from './animationHelpers.js';
 
-    'left strafe walking.fbx': new THREE.Vector3(-0.5, 0, 0),
-    'left strafe.fbx': new THREE.Vector3(-1, 0, 0),
-    'right strafe walking.fbx': new THREE.Vector3(0.5, 0, 0),
-    'right strafe.fbx': new THREE.Vector3(1, 0, 0),
-
-    'Fast Run.fbx': new THREE.Vector3(0, 0, -1),
-    'walking.fbx': new THREE.Vector3(0, 0, -0.5),
-
-    'running backwards.fbx': new THREE.Vector3(0, 0, 1),
-    'walking backwards.fbx': new THREE.Vector3(0, 0, 0.5),
-
-    'left strafe walking reverse.fbx': new THREE.Vector3(-Infinity, 0, 0),
-    'left strafe reverse.fbx': new THREE.Vector3(-Infinity, 0, 0),
-    'right strafe walking reverse.fbx': new THREE.Vector3(Infinity, 0, 0),
-    'right strafe reverse.fbx': new THREE.Vector3(Infinity, 0, 0),
-  },
-};
-const animationsDistanceMap = {
-  'idle.fbx': new THREE.Vector3(0, 0, 0),
-  'jump.fbx': new THREE.Vector3(0, 1, 0),
-
-  'left strafe walking.fbx': new THREE.Vector3(-0.5, 0, 0),
-  'left strafe.fbx': new THREE.Vector3(-1, 0, 0),
-  'right strafe walking.fbx': new THREE.Vector3(0.5, 0, 0),
-  'right strafe.fbx': new THREE.Vector3(1, 0, 0),
-
-  'Fast Run.fbx': new THREE.Vector3(0, 0, -1),
-  'walking.fbx': new THREE.Vector3(0, 0, -0.5),
-
-  'running backwards.fbx': new THREE.Vector3(0, 0, 1),
-  'walking backwards.fbx': new THREE.Vector3(0, 0, 0.5),
-
-  'left strafe walking reverse.fbx': new THREE.Vector3(-1, 0, 1).normalize().multiplyScalar(2),
-  'left strafe reverse.fbx': new THREE.Vector3(-1, 0, 1).normalize().multiplyScalar(3),
-  'right strafe walking reverse.fbx': new THREE.Vector3(1, 0, 1).normalize().multiplyScalar(2),
-  'right strafe reverse.fbx': new THREE.Vector3(1, 0, 1).normalize().multiplyScalar(3),
-  
-  'Crouch Idle.fbx': new THREE.Vector3(0, 0, 0),
-  'Sneaking Forward.fbx': new THREE.Vector3(0, 0, -crouchMagnitude),
-  'Sneaking Forward reverse.fbx': new THREE.Vector3(0, 0, crouchMagnitude),
-  'Crouched Sneaking Left.fbx': new THREE.Vector3(-crouchMagnitude, 0, 0),
-  'Crouched Sneaking Left reverse.fbx': new THREE.Vector3(-crouchMagnitude, 0, crouchMagnitude),
-  'Crouched Sneaking Right.fbx': new THREE.Vector3(crouchMagnitude, 0, 0),
-  'Crouched Sneaking Right reverse.fbx': new THREE.Vector3(crouchMagnitude, 0, crouchMagnitude),
-}; */
-const animationsAngleArrays = {
-  walk: [
-    {name: 'left strafe walking.fbx', angle: Math.PI/2},
-    {name: 'right strafe walking.fbx', angle: -Math.PI/2},
-
-    {name: 'walking.fbx', angle: 0},
-    {name: 'walking backwards.fbx', angle: Math.PI},
-
-    // {name: 'left strafe walking reverse.fbx', angle: Math.PI*3/4},
-    // {name: 'right strafe walking reverse.fbx', angle: -Math.PI*3/4},
-  ],
-  run: [
-    {name: 'left strafe.fbx', angle: Math.PI/2},
-    {name: 'right strafe.fbx', angle: -Math.PI/2},
-
-    {name: 'Fast Run.fbx', angle: 0},
-    {name: 'running backwards.fbx', angle: Math.PI},
-
-    // {name: 'left strafe reverse.fbx', angle: Math.PI*3/4},
-    // {name: 'right strafe reverse.fbx', angle: -Math.PI*3/4},
-  ],
-  crouch: [
-    {name: 'Crouched Sneaking Left.fbx', angle: Math.PI/2},
-    {name: 'Crouched Sneaking Right.fbx', angle: -Math.PI/2},
-    
-    {name: 'Sneaking Forward.fbx', angle: 0},
-    {name: 'Sneaking Forward reverse.fbx', angle: Math.PI},
-    
-    // {name: 'Crouched Sneaking Left reverse.fbx', angle: Math.PI*3/4},
-    // {name: 'Crouched Sneaking Right reverse.fbx', angle: -Math.PI*3/4},
-  ],
-};
-const animationsAngleArraysMirror = {
-  walk: [
-    {name: 'left strafe walking reverse.fbx', matchAngle: -Math.PI/2, angle: -Math.PI/2},
-    {name: 'right strafe walking reverse.fbx', matchAngle: Math.PI/2, angle: Math.PI/2},
-  ],
-  run: [
-    {name: 'left strafe reverse.fbx', matchAngle: -Math.PI/2, angle: -Math.PI/2},
-    {name: 'right strafe reverse.fbx', matchAngle: Math.PI/2, angle: Math.PI/2},
-  ],
-  crouch: [
-    {name: 'Crouched Sneaking Left reverse.fbx', matchAngle: -Math.PI/2, angle: -Math.PI/2},
-    {name: 'Crouched Sneaking Right reverse.fbx', matchAngle: Math.PI/2, angle: Math.PI/2},
-  ],
-};
-const animationsIdleArrays = {
-  reset: {name: 'reset.fbx'},
-  walk: {name: 'idle.fbx'},
-  run: {name: 'idle.fbx'},
-  crouch: {name: 'Crouch Idle.fbx'},
-};
-
-let animations;
-let animationStepIndices;
-let animationsBaseModel;
-let jumpAnimation;
-let floatAnimation;
-let useAnimations;
-let aimAnimations;
-let sitAnimations;
-let danceAnimations;
-let throwAnimations;
-// let crouchAnimations;
-let activateAnimations;
-let narutoRunAnimations;
-// let jumpAnimationSegments;
-// let chargeJump;
-// let standCharge;
-// let fallLoop;
-// let swordSideSlash;
-// let swordTopDownSlash;
-let hurtAnimations;
-const loadPromise = (async () => {
-  await Promise.resolve(); // wait for metaversefile to be defined
-  
-  await Promise.all([
-    (async () => {
-      const res = await fetch('/animations/animations.z');
-      const arrayBuffer = await res.arrayBuffer();
-      const uint8Array = new Uint8Array(arrayBuffer);
-      const animationsJson = zbdecode(uint8Array);
-      animations = animationsJson.animations
-        .map(a => THREE.AnimationClip.parse(a));
-      animationStepIndices = animationsJson.animationStepIndices;
-      animations.index = {};
-      for (const animation of animations) {
-        animations.index[animation.name] = animation;
-      }
-
-      /* const animationIndices = animationStepIndices.find(i => i.name === 'Fast Run.fbx');
-      for (let i = 0; i < animationIndices.leftFootYDeltas.length; i++) {
-        const mesh = new THREE.Mesh(new THREE.BoxBufferGeometry(0.02, 0.02, 0.02), new THREE.MeshBasicMaterial({color: 0xff0000}));
-        mesh.position.set(-30 + i * 0.1, 10 + animationIndices.leftFootYDeltas[i] * 10, -15);
-        mesh.updateMatrixWorld();
-        scene.add(mesh);
-      }
-      for (let i = 0; i < animationIndices.rightFootYDeltas.length; i++) {
-        const mesh = new THREE.Mesh(new THREE.BoxBufferGeometry(0.02, 0.02, 0.02), new THREE.MeshBasicMaterial({color: 0x0000ff}));
-        mesh.position.set(-30 + i * 0.1, 10 + animationIndices.rightFootYDeltas[i] * 10, -15);
-        mesh.updateMatrixWorld();
-        scene.add(mesh);
-      } */
-    })(),
-    (async () => {
-      const srcUrl = '/animations/animations-skeleton.glb';
-      
-      let o;
-      try {
-        o = await new Promise((accept, reject) => {
-          const {gltfLoader} = metaversefile.useLoaders();
-          gltfLoader.load(srcUrl, accept, function onprogress() {}, reject);
-        });
-      } catch(err) {
-        console.warn(err);
-      }
-      if (o) {
-        animationsBaseModel = o;
-      }
-    })(),
-  ]);
-
-  for (const k in animationsAngleArrays) {
-    const as = animationsAngleArrays[k];
-    for (const a of as) {
-      a.animation = animations.index[a.name];
-    }
-  }
-  for (const k in animationsAngleArraysMirror) {
-    const as = animationsAngleArraysMirror[k];
-    for (const a of as) {
-      a.animation = animations.index[a.name];
-    }
-  }
-  for (const k in animationsIdleArrays) {
-    animationsIdleArrays[k].animation = animations.index[animationsIdleArrays[k].name];
-  }
-
-  const _normalizeAnimationDurations = (animations, baseAnimation, factor = 1) => {
-    for (let i = 1; i < animations.length; i++) {
-      const animation = animations[i];
-      const oldDuration = animation.duration;
-      const newDuration = baseAnimation.duration;
-      for (const track of animation.tracks) {
-        const {times} = track;
-        for (let j = 0; j < times.length; j++) {
-          times[j] *= newDuration / oldDuration * factor;
-        }
-      }
-      animation.duration = newDuration * factor;
-    }
-  };
-  const walkingAnimations = [
-    `walking.fbx`,
-    `left strafe walking.fbx`,
-    `right strafe walking.fbx`,
-  ].map(name => animations.index[name]);
-  _normalizeAnimationDurations(walkingAnimations, walkingAnimations[0]);
-  const walkingBackwardAnimations = [
-    `walking backwards.fbx`,
-    `left strafe walking reverse.fbx`,
-    `right strafe walking reverse.fbx`,
-  ].map(name => animations.index[name]);
-  _normalizeAnimationDurations(walkingBackwardAnimations, walkingBackwardAnimations[0]);
-  const runningAnimations = [
-    `Fast Run.fbx`,
-    `left strafe.fbx`,
-    `right strafe.fbx`,
-  ].map(name => animations.index[name]);
-  _normalizeAnimationDurations(runningAnimations, runningAnimations[0]);
-  const runningBackwardAnimations = [
-    `running backwards.fbx`,
-    `left strafe reverse.fbx`,
-    `right strafe reverse.fbx`,
-  ].map(name => animations.index[name]);
-  _normalizeAnimationDurations(runningBackwardAnimations, runningBackwardAnimations[0]);
-  const crouchingForwardAnimations = [
-    `Sneaking Forward.fbx`,
-    `Crouched Sneaking Left.fbx`,
-    `Crouched Sneaking Right.fbx`,
-  ].map(name => animations.index[name]);
-  _normalizeAnimationDurations(crouchingForwardAnimations, crouchingForwardAnimations[0], 0.5);
-  const crouchingBackwardAnimations = [
-    `Sneaking Forward reverse.fbx`,
-    `Crouched Sneaking Left reverse.fbx`,
-    `Crouched Sneaking Right reverse.fbx`,
-  ].map(name => animations.index[name]);
-  _normalizeAnimationDurations(crouchingBackwardAnimations, crouchingBackwardAnimations[0], 0.5);
-  for (const animation of animations) {
-    decorateAnimation(animation);
-  }
-
-  /* jumpAnimationSegments = {
-    chargeJump: animations.find(a => a.isChargeJump),
-    chargeJumpFall: animations.find(a => a.isChargeJumpFall),
-    isFallLoop: animations.find(a => a.isFallLoop),
-    isLanding: animations.find(a => a.isLanding)
-  }; */
-
-  // chargeJump = animations.find(a => a.isChargeJump);
-  // standCharge = animations.find(a => a.isStandCharge);
-  // fallLoop = animations.find(a => a.isFallLoop);
-  // swordSideSlash = animations.find(a => a.isSwordSideSlash);
-  // swordTopDownSlash = animations.find(a => a.isSwordTopDownSlash)
-
-  function mergeAnimations(a, b) {
-    const o = {};
-    for (const k in a) {
-      o[k] = a[k];
-    }
-    for (const k in b) {
-      o[k] = b[k];
-    }
-    return o;
-  }
-
-  jumpAnimation = animations.find(a => a.isJump);
-  // sittingAnimation = animations.find(a => a.isSitting);
-  floatAnimation = animations.find(a => a.isFloat);
-  // rifleAnimation = animations.find(a => a.isRifle);
-  // hitAnimation = animations.find(a => a.isHit);
-  aimAnimations = {
-    swordSideIdle: animations.index['sword_idle_side.fbx'],
-    swordSideIdleStatic: animations.index['sword_idle_side_static.fbx'],
-    swordSideSlash: animations.index['sword_side_slash.fbx'],
-    swordSideSlashStep: animations.index['sword_side_slash_step.fbx'],
-    swordTopDownSlash: animations.index['sword_topdown_slash.fbx'],
-    swordTopDownSlashStep: animations.index['sword_topdown_slash_step.fbx'],
-    swordUndraw: animations.index['sword_undraw.fbx'],
-  };
-  useAnimations = mergeAnimations({
-    combo: animations.find(a => a.isCombo),
-    slash: animations.find(a => a.isSlash),
-    rifle: animations.find(a => a.isRifle),
-    pistol: animations.find(a => a.isPistol),
-    magic: animations.find(a => a.isMagic),
-    eat: animations.find(a => a.isEating),
-    drink: animations.find(a => a.isDrinking),
-    throw: animations.find(a => a.isThrow),
-    bowDraw: animations.find(a => a.isBowDraw),
-    bowIdle: animations.find(a => a.isBowIdle),
-    bowLoose: animations.find(a => a.isBowLoose),
-  }, aimAnimations);
-  sitAnimations = {
-    chair: animations.find(a => a.isSitting),
-    saddle: animations.find(a => a.isSitting),
-    stand: animations.find(a => a.isSkateboarding),
-  };
-  danceAnimations = {
-    dansu: animations.find(a => a.isDancing),
-    powerup: animations.find(a => a.isPowerUp),
-  };
-  throwAnimations = {
-    throw: animations.find(a => a.isThrow),
-  };
-  /* crouchAnimations = {
-    crouch: animations.find(a => a.isCrouch),
-  }; */
-  activateAnimations = {
-    grab_forward: {animation:animations.index['grab_forward.fbx'], speedFactor: 1.2},
-    grab_down: {animation:animations.index['grab_down.fbx'], speedFactor: 1.7},
-    grab_up: {animation:animations.index['grab_up.fbx'], speedFactor: 1.2},
-    grab_left: {animation:animations.index['grab_left.fbx'], speedFactor: 1.2},
-    grab_right: {animation:animations.index['grab_right.fbx'], speedFactor: 1.2},
-  };
-  narutoRunAnimations = {
-    narutoRun: animations.find(a => a.isNarutoRun),
-  };
-  hurtAnimations = {
-    pain_back: animations.index['pain_back.fbx'],
-    pain_arch: animations.index['pain_arch.fbx'],
-  };
-  {
-    const down10QuaternionArray = new THREE.Quaternion()
-      .setFromAxisAngle(new THREE.Vector3(1, 0, 0), Math.PI*0.1)
-      .toArray();
-    [
-      'mixamorigSpine1.quaternion',
-      'mixamorigSpine2.quaternion',
-    ].forEach(k => {
-      narutoRunAnimations.narutoRun.interpolants[k].evaluate = t => down10QuaternionArray;
-    });
-  }
-})().catch(err => {
-  console.log('load avatar animations error', err);
-});
-
-const _localizeMatrixWorld = bone => {
-  bone.matrix.copy(bone.matrixWorld);
-  if (bone.parent) {
-    bone.matrix.premultiply(bone.parent.matrixWorld.clone().invert());
-  }
-  bone.matrix.decompose(bone.position, bone.quaternion, bone.scale);
-
-  for (let i = 0; i < bone.children.length; i++) {
-    _localizeMatrixWorld(bone.children[i]);
-  }
-};
-const _findBoneDeep = (bones, boneName) => {
-  for (let i = 0; i < bones.length; i++) {
-    const bone = bones[i];
-    if (bone.name === boneName) {
-      return bone;
-    } else {
-      const deepBone = _findBoneDeep(bone.children, boneName);
-      if (deepBone) {
-        return deepBone;
-      }
-    }
-  }
-  return null;
-};
-/* const copySkeleton = (src, dst) => {
-  for (let i = 0; i < src.bones.length; i++) {
-    const srcBone = src.bones[i];
-    const dstBone = _findBoneDeep(dst.bones, srcBone.name);
-    dstBone.matrixWorld.copy(srcBone.matrixWorld);
-  }
-
-  // const armature = dst.bones[0].parent;
-  // _localizeMatrixWorld(armature);
-
-  dst.calculateInverses();
-}; */
 
 const cubeGeometry = new THREE.BoxGeometry(1, 1, 1);
 const srcCubeGeometries = {};
@@ -561,14 +131,14 @@ const debugMeshMaterial = new THREE.MeshNormalMaterial({
   transparent: true,
   depthTest: false,
 });
-const _makeDebugMesh = () => {
+const _makeDebugMesh = (avatar) => {
   const baseScale = 0.01;
   const fingerScale = 0.5;
   const _makeCubeMesh = (name, scale = 1) => {
     let srcGeometry = srcCubeGeometries[scale];
     if (!srcGeometry) {
       srcGeometry = cubeGeometry.clone()
-        .applyMatrix4(localMatrix.makeScale(scale * baseScale, scale * baseScale, 1))
+        .applyMatrix4(avatar.localMatrix.makeScale(scale * baseScale, scale * baseScale, 1))
       srcCubeGeometries[scale] = srcGeometry;
     }
     const geometry = srcGeometry//.clone();
@@ -725,10 +295,10 @@ const _makeDebugMesh = () => {
       const meshBone = attributes[k];
 
       modelBone.forwardQuaternion = new THREE.Quaternion().setFromRotationMatrix(
-        localMatrix.lookAt(
-          localVector.set(0, 0, 0),
+        avatar.localMatrix.lookAt(
+          avatar.localVector.set(0, 0, 0),
           modelBone.position,
-          localVector2.set(0, 1, 0)
+          avatar.localVector2.set(0, 1, 0)
         )
       );
       modelBoneToMeshBoneMap.set(modelBone, meshBone);
@@ -742,23 +312,23 @@ const _makeDebugMesh = () => {
       (modelBone.parent ?
         modelBone.parent.matrixWorld
       :
-        localMatrix.identity()
-      ).decompose(localVector, localQuaternion, localVector2);
+        avatar.localMatrix.identity()
+      ).decompose(avatar.localVector, avatar.localQuaternion, avatar.localVector2);
 
-      localVector.add(
-        localVector3.set(0, 0, -meshBone.mesh.scale.z*0.5)
+      avatar.localVector.add(
+        avatar.localVector3.set(0, 0, -meshBone.mesh.scale.z*0.5)
           .applyQuaternion(modelBone.forwardQuaternion)
-          .applyQuaternion(localQuaternion)
+          .applyQuaternion(avatar.localQuaternion)
       );
 
-      localQuaternion.multiply(
+      avatar.localQuaternion.multiply(
         modelBone.forwardQuaternion
       );
 
-      meshBone.matrixWorld.compose(localVector, localQuaternion, localVector2);
+      meshBone.matrixWorld.compose(avatar.localVector, avatar.localQuaternion, avatar.localVector2);
       meshBone.matrix.copy(meshBone.matrixWorld);
       if (meshBone.parent) {
-        meshBone.matrix.premultiply(localMatrix.copy(meshBone.parent.matrixWorld).invert());
+        meshBone.matrix.premultiply(avatar.localMatrix.copy(meshBone.parent.matrixWorld).invert());
       }
       meshBone.matrix.decompose(meshBone.position, meshBone.quaternion, meshBone.scale);
     }
@@ -806,112 +376,7 @@ const _makeDebugMesh = () => {
   };
   return mesh;
 };
-const _findArmature = bone => {
-  for (;; bone = bone.parent) {
-    if (!bone.isBone) {
-      return bone;
-    }
-  }
-  return null; // can't happen
-};
 
-/* const _exportBone = bone => {
-  return [bone.name, bone.position.toArray().concat(bone.quaternion.toArray()).concat(bone.scale.toArray()), bone.children.map(b => _exportBone(b))];
-};
-const _exportSkeleton = skeleton => {
-  const hips = _findHips(skeleton);
-  const armature = _findArmature(hips);
-  return JSON.stringify(_exportBone(armature));
-};
-const _importObject = (b, Cons, ChildCons) => {
-  const [name, array, children] = b;
-  const bone = new Cons();
-  bone.name = name;
-  bone.position.fromArray(array, 0);
-  bone.quaternion.fromArray(array, 3);
-  bone.scale.fromArray(array, 3+4);
-  for (let i = 0; i < children.length; i++) {
-    bone.add(_importObject(children[i], ChildCons, ChildCons));
-  }
-  return bone;
-};
-const _importArmature = b => _importObject(b, THREE.Object3D, THREE.Bone);
-const _importSkeleton = s => {
-  const armature = _importArmature(JSON.parse(s));
-  return new THREE.Skeleton(armature.children);
-}; */
-
-class AnimationMapping {
-  constructor(animationTrackName, boneName, isTop, isPosition) {
-    this.animationTrackName = animationTrackName;
-    this.boneName = boneName;
-    this.isTop = isTop;
-    this.isPosition = isPosition;
-  }
-  clone() {
-    return new AnimationMapping(this.animationTrackName, this.boneName, this.isTop, this.isPosition);
-  }
-}
-const _getLerpFn = isPosition => isPosition ? THREE.Vector3.prototype.lerp : THREE.Quaternion.prototype.slerp;
-const animationMappingConfig = [
-  new AnimationMapping('mixamorigHips.position', 'Hips', false, true),
-  new AnimationMapping('mixamorigHips.quaternion', 'Hips', false, false),
-  new AnimationMapping('mixamorigSpine.quaternion', 'Spine', true, false),
-  new AnimationMapping('mixamorigSpine1.quaternion', 'Chest', true, false),
-  new AnimationMapping('mixamorigSpine2.quaternion', 'UpperChest', true, false),
-  new AnimationMapping('mixamorigNeck.quaternion', 'Neck', true, false),
-  new AnimationMapping('mixamorigHead.quaternion', 'Head', true, false),
-
-  new AnimationMapping('mixamorigLeftShoulder.quaternion', 'Left_shoulder', true, false),
-  new AnimationMapping('mixamorigLeftArm.quaternion', 'Left_arm', true, false),
-  new AnimationMapping('mixamorigLeftForeArm.quaternion', 'Left_elbow', true, false),
-  new AnimationMapping('mixamorigLeftHand.quaternion', 'Left_wrist', true, false),
-  new AnimationMapping('mixamorigLeftHandMiddle1.quaternion', 'Left_middleFinger1', true, false),
-  new AnimationMapping('mixamorigLeftHandMiddle2.quaternion', 'Left_middleFinger2', true, false),
-  new AnimationMapping('mixamorigLeftHandMiddle3.quaternion', 'Left_middleFinger3', true, false),
-  new AnimationMapping('mixamorigLeftHandThumb1.quaternion', 'Left_thumb0', true, false),
-  new AnimationMapping('mixamorigLeftHandThumb2.quaternion', 'Left_thumb1', true, false),
-  new AnimationMapping('mixamorigLeftHandThumb3.quaternion', 'Left_thumb2', true, false),
-  new AnimationMapping('mixamorigLeftHandIndex1.quaternion', 'Left_indexFinger1', true, false),
-  new AnimationMapping('mixamorigLeftHandIndex2.quaternion', 'Left_indexFinger2', true, false),
-  new AnimationMapping('mixamorigLeftHandIndex3.quaternion', 'Left_indexFinger3', true, false),
-  new AnimationMapping('mixamorigLeftHandRing1.quaternion', 'Left_ringFinger1', true, false),
-  new AnimationMapping('mixamorigLeftHandRing2.quaternion', 'Left_ringFinger2', true, false),
-  new AnimationMapping('mixamorigLeftHandRing3.quaternion', 'Left_ringFinger3', true, false),
-  new AnimationMapping('mixamorigLeftHandPinky1.quaternion', 'Left_littleFinger1', true, false),
-  new AnimationMapping('mixamorigLeftHandPinky2.quaternion', 'Left_littleFinger2', true, false),
-  new AnimationMapping('mixamorigLeftHandPinky3.quaternion', 'Left_littleFinger3', true, false),
-
-  new AnimationMapping('mixamorigRightShoulder.quaternion', 'Right_shoulder', true, false),
-  new AnimationMapping('mixamorigRightArm.quaternion', 'Right_arm', true, false),
-  new AnimationMapping('mixamorigRightForeArm.quaternion', 'Right_elbow', true, false),
-  new AnimationMapping('mixamorigRightHand.quaternion', 'Right_wrist', true, false),
-  new AnimationMapping('mixamorigRightHandMiddle1.quaternion', 'Right_middleFinger1', true, false),
-  new AnimationMapping('mixamorigRightHandMiddle2.quaternion', 'Right_middleFinger2', true, false),
-  new AnimationMapping('mixamorigRightHandMiddle3.quaternion', 'Right_middleFinger3', true, false),
-  new AnimationMapping('mixamorigRightHandThumb1.quaternion', 'Right_thumb0', true, false),
-  new AnimationMapping('mixamorigRightHandThumb2.quaternion', 'Right_thumb1', true, false),
-  new AnimationMapping('mixamorigRightHandThumb3.quaternion', 'Right_thumb2', true, false),
-  new AnimationMapping('mixamorigRightHandIndex1.quaternion', 'Right_indexFinger1', true, false),
-  new AnimationMapping('mixamorigRightHandIndex2.quaternion', 'Right_indexFinger2', true, false),
-  new AnimationMapping('mixamorigRightHandIndex3.quaternion', 'Right_indexFinger3', true, false),
-  new AnimationMapping('mixamorigRightHandRing1.quaternion', 'Right_ringFinger1', true, false),
-  new AnimationMapping('mixamorigRightHandRing2.quaternion', 'Right_ringFinger2', true, false),
-  new AnimationMapping('mixamorigRightHandRing3.quaternion', 'Right_ringFinger3', true, false),
-  new AnimationMapping('mixamorigRightHandPinky1.quaternion', 'Right_littleFinger1', true, false),
-  new AnimationMapping('mixamorigRightHandPinky2.quaternion', 'Right_littleFinger2', true, false),
-  new AnimationMapping('mixamorigRightHandPinky3.quaternion', 'Right_littleFinger3', true, false),
-
-  new AnimationMapping('mixamorigRightUpLeg.quaternion', 'Right_leg', false, false),
-  new AnimationMapping('mixamorigRightLeg.quaternion', 'Right_knee', false, false),
-  new AnimationMapping('mixamorigRightFoot.quaternion', 'Right_ankle', false, false),
-  new AnimationMapping('mixamorigRightToeBase.quaternion', 'Right_toe', false, false),
-
-  new AnimationMapping('mixamorigLeftUpLeg.quaternion', 'Left_leg', false, false),
-  new AnimationMapping('mixamorigLeftLeg.quaternion', 'Left_knee', false, false),
-  new AnimationMapping('mixamorigLeftFoot.quaternion', 'Left_ankle', false, false),
-  new AnimationMapping('mixamorigLeftToeBase.quaternion', 'Left_toe', false, false),
-];
 const _clearXZ = (dst, isPosition) => {
   if (isPosition) {
     dst.x = 0;
@@ -919,201 +384,6 @@ const _clearXZ = (dst, isPosition) => {
   }
 };
 
-class Blinker {
-  constructor() {
-    this.mode = 'ready';
-    this.waitTime = 0;
-    this.lastTimestamp = 0;
-  }
-  update(now) {
-    const _setOpen = () => {
-      this.mode = 'open';
-      this.waitTime = (0.5 + 0.5 * Math.random()) * 3000;
-      this.lastTimestamp = now;
-    };
-
-    switch (this.mode) {
-      case 'ready': {
-        _setOpen();
-        return 0;
-      }
-      case 'open': {
-        const timeDiff = now - this.lastTimestamp;
-        if (timeDiff > this.waitTime) {
-          this.mode = 'closing';
-          this.waitTime = 100;
-          this.lastTimestamp = now;
-        }
-        return 0;
-      }
-      case 'closing': {
-        const f = Math.min(Math.max((now - this.lastTimestamp) / this.waitTime, 0), 1);
-        if (f < 1) {
-          return f;
-        } else {
-          this.mode = 'opening';
-          this.waitTime = 100;
-          this.lastTimestamp = now;
-          return 1;
-        }
-      }
-      case 'opening': {
-        const f = Math.min(Math.max((now - this.lastTimestamp) / this.waitTime, 0), 1);
-        if (f < 1) {
-          return 1 - f;
-        } else {
-          _setOpen();
-          return 0;
-        }
-      }
-    }
-  }
-}
-class Nodder {
-  constructor() {
-
-  }
-  update() {
-
-  }
-}
-// const g = new THREE.BoxBufferGeometry(0.05, 0.05, 0.05);
-// const m = new THREE.MeshBasicMaterial({ color: 0xFF00FF });
-// const testMesh = new THREE.Mesh(g, m);
-// scene.add(testMesh);
-class Looker {
-  constructor(avatar) {
-    this.avatar = avatar;
-
-    this.mode = 'ready';
-    this.startTarget = new THREE.Vector3();
-    this.endTarget = new THREE.Vector3();
-    this.waitTime = 0;
-    this.lastTimestamp = 0;
-
-    this._target = new THREE.Vector3();
-  }
-  // returns the world space eye target
-  update(now) {
-    const _getEndTargetRandom = target => {
-      const root = this.avatar.modelBoneOutputs['Root'];
-      const eyePosition = getEyePosition(this.avatar.modelBones);
-      return target.copy(eyePosition)
-        .add(
-          localVector.set(0, 0, 1.5 + 3 * Math.random())
-            .applyQuaternion(localQuaternion.setFromRotationMatrix(root.matrixWorld))
-        )
-        .add(
-          localVector.set(-0.5+Math.random(), (-0.5+Math.random()) * 0.3, -0.5+Math.random())
-            .normalize()
-            // .multiplyScalar(1)
-        );
-    };
-    const _getEndTargetForward = target => {
-      const root = this.avatar.modelBoneOutputs['Root'];
-      const eyePosition = getEyePosition(this.avatar.modelBones);
-      return target.copy(eyePosition)
-        .add(
-          localVector.set(0, 0, 2)
-            .applyQuaternion(localQuaternion.setFromRotationMatrix(root.matrixWorld))
-        );
-    };
-    const _startMove = () => {
-      this.mode = 'moving';
-      // const head = this.avatar.modelBoneOutputs['Head'];
-      // const root = this.avatar.modelBoneOutputs['Root'];
-      this.startTarget.copy(this.endTarget);
-      _getEndTargetRandom(this.endTarget);
-      this.waitTime = 100;
-      this.lastTimestamp = now;
-    };
-    const _startDelay = () => {
-      this.mode = 'delay';
-      this.waitTime = Math.random() * 2000;
-      this.lastTimestamp = now;
-    };
-    const _startWaiting = () => {
-      this.mode = 'waiting';
-      this.waitTime = Math.random() * 3000;
-      this.lastTimestamp = now;
-    };
-    const _isSpeedTooFast = () => this.avatar.velocity.length() > 0.5;
-    const _isPointTooClose = () => {
-      const root = this.avatar.modelBoneOutputs['Root'];
-      // const head = this.avatar.modelBoneOutputs['Head'];
-      localVector.set(0, 0, 1)
-        .applyQuaternion(localQuaternion.setFromRotationMatrix(root.matrixWorld));
-      localVector2.setFromMatrixPosition(root.matrixWorld);
-      localPlane.setFromNormalAndCoplanarPoint(
-        localVector,
-        localVector2
-      );
-      const distance = localPlane.distanceToPoint(this.endTarget);
-      return distance < 1;
-    };
-
-    // console.log('got mode', this.mode, this.waitTime);
-    
-    if (_isSpeedTooFast()) {
-      _getEndTargetForward(this.endTarget);
-      // this.startTarget.copy(this.endTarget);
-      _startDelay();
-      return null;
-    } else if (_isPointTooClose()) {
-      _getEndTargetForward(this.endTarget);
-      // this.startTarget.copy(this.endTarget);
-      _startDelay();
-      return null;
-    } else {
-      switch (this.mode) {
-        case 'ready': {
-          _startMove();
-          return this.startTarget;
-        }
-        case 'delay': {
-          const timeDiff = now - this.lastTimestamp;
-          if (timeDiff > this.waitTime) {
-            _startMove();
-            return this.startTarget;
-          } else {
-            return null;
-          }
-        }
-        case 'moving': {
-          const timeDiff = now - this.lastTimestamp;
-          const f = Math.min(Math.max(timeDiff / this.waitTime, 0), 1);
-          // console.log('got time diff', timeDiff, this.waitTime, f);
-          const target = this._target.copy(this.startTarget)
-            .lerp(this.endTarget, f);
-          // _setTarget(target);
-
-          if (f >= 1) {
-            _startWaiting();
-          }
-
-          return target;
-        }
-        case 'waiting': {
-          const f = Math.min(Math.max((now - this.lastTimestamp) / this.waitTime, 0), 1);
-          if (f >= 1) {
-            _startMove();
-            return this.startTarget;
-          } else {
-            return this.endTarget;
-          }
-        }
-      }
-    }
-  }
-}
-class Emoter {
-  constructor() {
-    
-  }
-  update() {
-    
-  }
-}
 
 class Avatar {
 	constructor(object, options = {}) {
@@ -1163,6 +433,19 @@ class Avatar {
 
     this.vrmExtension = object?.parser?.json?.extensions?.VRM;
     this.firstPersonCurves = getFirstPersonCurves(this.vrmExtension);
+
+    this.localVector = new THREE.Vector3();
+    this.localVector2 = new THREE.Vector3();
+    this.localVector3 = new THREE.Vector3();
+
+    this.localQuaternion = new THREE.Quaternion();
+    this.localQuaternion2 = new THREE.Quaternion();
+
+    this.localEuler = new THREE.Euler(0, 0, 0, 'YXZ');
+    this.localEulerHmd = new THREE.Euler(0, 0, 0, 'YXZ');
+    this.localMatrix = new THREE.Matrix4();
+    this.localMatrix2 = new THREE.Matrix4();
+    this.localPlane = new THREE.Plane();
 
     const {
       skinnedMeshes,
@@ -1764,7 +1047,7 @@ class Avatar {
           _recurseBoneAttachments(child);
         } else {
           child.matrix
-            .premultiply(localMatrix.compose(localVector.set(0, 0, 0), new THREE.Quaternion().setFromAxisAngle(new THREE.Vector3(0, 1, 0), Math.PI), localVector2.set(1, 1, 1)))
+            .premultiply(this.localMatrix.compose(this.localVector.set(0, 0, 0), new THREE.Quaternion().setFromAxisAngle(new THREE.Vector3(0, 1, 0), Math.PI), this.localVector2.set(1, 1, 1)))
             .decompose(child.position, child.quaternion, child.scale);
         }
       }
@@ -2055,15 +1338,15 @@ class Avatar {
     return this.legsManager.enabled;
   }
   getAngle() {
-    localEuler.setFromRotationMatrix(
-      localMatrix.lookAt(
-        localVector.set(0, 0, 0),
+    this.localEuler.setFromRotationMatrix(
+      this.localMatrix.lookAt(
+        this.localVector.set(0, 0, 0),
         this.direction,
-        localVector2.set(0, 1, 0)
+        this.localVector2.set(0, 1, 0)
       ),
       'YXZ'
     );
-    return localEuler.y;
+    return this.localEuler.y;
   }
   async setQuality(quality) {
 
@@ -2104,27 +1387,29 @@ class Avatar {
   update(timestamp, timeDiff) {
     const now = timestamp;
     const timeDiffS = timeDiff / 1000;
-    const currentSpeed = localVector.set(this.velocity.x, 0, this.velocity.z).length();
+    const currentSpeed = this.localVector.set(this.velocity.x, 0, this.velocity.z).length();
 
-    const idleWalkFactor = Math.min(Math.max((currentSpeed - idleFactorSpeed) / (walkFactorSpeed - idleFactorSpeed), 0), 1);
-    const walkRunFactor = Math.min(Math.max((currentSpeed - walkFactorSpeed) / (runFactorSpeed - walkFactorSpeed), 0), 1);
-    const crouchFactor = Math.min(Math.max(1 - (this.crouchTime / crouchMaxTime), 0), 1);
+    const moveFactors = {};
+    moveFactors.idleWalkFactor = Math.min(Math.max((currentSpeed - idleFactorSpeed) / (walkFactorSpeed - idleFactorSpeed), 0), 1);
+    moveFactors.walkRunFactor = Math.min(Math.max((currentSpeed - walkFactorSpeed) / (runFactorSpeed - walkFactorSpeed), 0), 1);
+    moveFactors.crouchFactor = Math.min(Math.max(1 - (this.crouchTime / crouchMaxTime), 0), 1);
     // console.log('current speed', currentSpeed, idleWalkFactor, walkRunFactor);
 
-    const _updatePosition = () => {
+    const _updateHmdPosition = () => {
       const currentPosition = this.inputs.hmd.position;
       const currentQuaternion = this.inputs.hmd.quaternion;
+      // console.log("INPUTS HMD POS AND QUAT", currentPosition, currentQuaternion)
       
-      const positionDiff = localVector.copy(this.lastPosition)
+      const positionDiff = this.localVector.copy(this.lastPosition)
         .sub(currentPosition)
         .divideScalar(timeDiffS)
         .multiplyScalar(0.1);
-      localEuler.setFromQuaternion(currentQuaternion, 'YXZ');
-      localEuler.x = 0;
-      localEuler.z = 0;
-      localEuler.y += Math.PI;
-      localEuler2.set(-localEuler.x, -localEuler.y, -localEuler.z, localEuler.order);
-      positionDiff.applyEuler(localEuler2);
+      this.localEuler.setFromQuaternion(currentQuaternion, 'YXZ');
+      this.localEuler.x = 0;
+      this.localEuler.z = 0;
+      this.localEuler.y += Math.PI;
+      this.localEulerHmd.set(-this.localEuler.x, -this.localEuler.y, -this.localEuler.z, this.localEuler.order);
+      positionDiff.applyEuler(this.localEulerHmd);
       this.velocity.copy(positionDiff);
       this.lastPosition.copy(currentPosition);
       this.direction.copy(positionDiff).normalize();
@@ -2133,764 +1418,6 @@ class Avatar {
         this.lastMoveTime = now;
       }
     };
-    _updatePosition();
-    
-    const _applyAnimation = () => {
-      // const runSpeed = 0.5;
-      const angle = this.getAngle();
-      const timeSeconds = now/1000;
-      
-      /* const _getAnimationKey = crouchState => {
-        if (crouchState) {
-          return 'crouch';
-        } else {
-          if (currentSpeed >= runSpeed) {
-            return 'run';
-          } else {
-            return 'walk';
-          }
-        }
-      }; */
-      const _getMirrorAnimationAngles = (animationAngles, key) => {
-        const animations = animationAngles.map(({animation}) => animation);
-        const animationAngleArrayMirror = animationsAngleArraysMirror[key];
-        
-        const backwardIndex = animations.findIndex(a => a.isBackward);
-        if (backwardIndex !== -1) {
-          // const backwardAnimationAngle = animationAngles[backwardIndex];
-          // const angleToBackwardAnimation = Math.abs(angleDifference(angle, backwardAnimationAngle.angle));
-          // if (angleToBackwardAnimation < Math.PI * 0.3) {
-            const sideIndex = backwardIndex === 0 ? 1 : 0;
-            const wrongAngle = animationAngles[sideIndex].angle;
-            const newAnimationAngle = animationAngleArrayMirror.find(animationAngle => animationAngle.matchAngle === wrongAngle);
-            animationAngles = animationAngles.slice();
-            animationAngles[sideIndex] = newAnimationAngle;
-            // animations[sideIndex] = newAnimationAngle.animation;
-            // return {
-              // return animationAngles;
-              // angleToBackwardAnimation,
-            // };
-          // }
-        }
-        // return {
-          return animationAngles;
-          // angleToBackwardAnimation: Infinity,
-        // ;
-      };
-      const _getAngleToBackwardAnimation = animationAngles => {
-        const animations = animationAngles.map(({animation}) => animation);
-        
-        const backwardIndex = animations.findIndex(a => a.isBackward);
-        if (backwardIndex !== -1) {
-          const backwardAnimationAngle = animationAngles[backwardIndex];
-          const angleToBackwardAnimation = Math.abs(angleDifference(angle, backwardAnimationAngle.angle));
-          return angleToBackwardAnimation;
-        } else {
-          return Infinity;
-        }
-      };
-      const _getIdleAnimation = key => animationsIdleArrays[key].animation;
-      /* const _getIdleAnimation = key => {
-        if (key === 'walk' || key === 'run') {
-          const name = animationsIdleArrays[key].name;
-          return this.retargetedAnimations.find(a => a.name === name);
-        } else {
-          return animationsIdleArrays[key].animation;
-        }
-      }; */
-      const _get7wayBlend = (
-        horizontalWalkAnimationAngles,
-        horizontalWalkAnimationAnglesMirror,
-        horizontalRunAnimationAngles,
-        horizontalRunAnimationAnglesMirror,
-        idleAnimation,
-        // mirrorFactor,
-        // angleFactor,
-        // walkRunFactor,
-        // idleWalkFactor,
-        k,
-        lerpFn,
-        isPosition,
-        target
-      ) => {
-        // WALK
-        // normal horizontal walk blend
-        {
-          const t1 = timeSeconds % horizontalWalkAnimationAngles[0].animation.duration;
-          const src1 = horizontalWalkAnimationAngles[0].animation.interpolants[k];
-          const v1 = src1.evaluate(t1);
-
-          const t2 = timeSeconds % horizontalWalkAnimationAngles[1].animation.duration;
-          const src2 = horizontalWalkAnimationAngles[1].animation.interpolants[k];
-          const v2 = src2.evaluate(t2);
-          
-          lerpFn
-            .call(
-              localQuaternion3.fromArray(v2),
-              localQuaternion4.fromArray(v1),
-              angleFactor
-            );
-        }
-          
-        // mirror horizontal blend (backwards walk)
-        {
-          const t1 = timeSeconds % horizontalWalkAnimationAnglesMirror[0].animation.duration;
-          const src1 = horizontalWalkAnimationAnglesMirror[0].animation.interpolants[k];
-          const v1 = src1.evaluate(t1);
-
-          const t2 = timeSeconds % horizontalWalkAnimationAnglesMirror[1].animation.duration;
-          const src2 = horizontalWalkAnimationAnglesMirror[1].animation.interpolants[k];
-          const v2 = src2.evaluate(t2);
-
-          lerpFn
-            .call(
-              localQuaternion4.fromArray(v2),
-              localQuaternion5.fromArray(v1),
-              angleFactor
-            );
-        }
-
-        // blend mirrors together to get a smooth walk
-        lerpFn
-          .call(
-            localQuaternion5.copy(localQuaternion3), // Result is in localQuaternion5
-            localQuaternion4,
-            mirrorFactor
-          );
-
-        // RUN
-        // normal horizontal run blend
-        {
-          const t1 = timeSeconds % horizontalRunAnimationAngles[0].animation.duration;
-          const src1 = horizontalRunAnimationAngles[0].animation.interpolants[k];
-          const v1 = src1.evaluate(t1);
-
-          const t2 = timeSeconds % horizontalRunAnimationAngles[1].animation.duration;
-          const src2 = horizontalRunAnimationAngles[1].animation.interpolants[k];
-          const v2 = src2.evaluate(t2);
-          
-          lerpFn
-            .call(
-              localQuaternion3.fromArray(v2),
-              localQuaternion4.fromArray(v1),
-              angleFactor
-            );
-        }
-          
-        // mirror horizontal blend (backwards run)
-        {
-          const t1 = timeSeconds % horizontalRunAnimationAnglesMirror[0].animation.duration;
-          const src1 = horizontalRunAnimationAnglesMirror[0].animation.interpolants[k];
-          const v1 = src1.evaluate(t1);
-
-          const t2 = timeSeconds % horizontalRunAnimationAnglesMirror[1].animation.duration;
-          const src2 = horizontalRunAnimationAnglesMirror[1].animation.interpolants[k];
-          const v2 = src2.evaluate(t2);
-
-          lerpFn
-            .call(
-              localQuaternion4.fromArray(v2),
-              localQuaternion6.fromArray(v1),
-              angleFactor
-            );
-        }
-
-        // blend mirrors together to get a smooth run
-        lerpFn
-          .call(
-            localQuaternion6.copy(localQuaternion3), // Result is in localQuaternion6
-            localQuaternion4,
-            mirrorFactor
-          );
-
-        // Blend walk/run
-        lerpFn
-          .call(
-            localQuaternion4.copy(localQuaternion5), // Result is in localQuaternion4
-            localQuaternion6,
-            walkRunFactor
-          );
-
-        // blend the smooth walk/run with idle
-        {
-          const timeSinceLastMove = now - this.lastMoveTime;
-          const timeSinceLastMoveSeconds = timeSinceLastMove / 1000;
-          const t3 = timeSinceLastMoveSeconds % idleAnimation.duration;
-          const src3 = idleAnimation.interpolants[k];
-          const v3 = src3.evaluate(t3);
-
-          target.fromArray(v3);
-          if (isPosition) {
-            // target.x = 0;
-            // target.z = 0;
-
-            localQuaternion4.x = 0;
-            localQuaternion4.z = 0;
-          }
-
-          lerpFn
-            .call(
-              target,
-              localQuaternion4,
-              idleWalkFactor
-            );
-        }
-      };
-      
-      // stand
-      // const key = _getAnimationKey(false);
-      const keyWalkAnimationAngles = getClosest2AnimationAngles('walk', angle);
-      const keyWalkAnimationAnglesMirror = _getMirrorAnimationAngles(keyWalkAnimationAngles, 'walk');
-
-      const keyRunAnimationAngles = getClosest2AnimationAngles('run', angle);
-      const keyRunAnimationAnglesMirror = _getMirrorAnimationAngles(keyRunAnimationAngles, 'run');
-      
-      const idleAnimation = _getIdleAnimation('walk');
-
-      /* // walk sound effect
-      {
-        const soundManager = metaversefile.useSoundManager();
-        const currAniTime = timeSeconds % idleAnimation.duration;
-
-        if (currentSpeed > 0.1) {
-          if (key === 'walk') {
-            if (currAniTime > 0.26 && currAniTime < 0.4)
-              soundManager.playStepSound(1);
-            if (currAniTime > 0.76 && currAniTime < 0.9)
-              soundManager.playStepSound(2);
-            if (currAniTime > 1.26 && currAniTime < 1.4)
-              soundManager.playStepSound(3);
-            if (currAniTime > 1.76 && currAniTime < 1.9)
-              soundManager.playStepSound(4);
-            if (currAniTime > 2.26 && currAniTime < 2.5)
-              soundManager.playStepSound(5);
-          }
-          if (key === 'run') {
-            if (currAniTime > 0.16 && currAniTime < 0.3)
-              soundManager.playStepSound(1);
-            if (currAniTime > 0.43 && currAniTime < 0.45)
-              soundManager.playStepSound(2);
-            if (currAniTime > 0.693 && currAniTime < 0.8)
-              soundManager.playStepSound(3);
-            if (currAniTime > 0.963 && currAniTime < 1.1)
-              soundManager.playStepSound(4);
-            if (currAniTime > 1.226 && currAniTime < 1.3)
-              soundManager.playStepSound(5);
-            if (currAniTime > 1.496 && currAniTime < 1.6)
-              soundManager.playStepSound(6);
-            if (currAniTime > 1.759 && currAniTime < 1.9)
-              soundManager.playStepSound(7);
-            if (currAniTime > 2.029 && currAniTime < 2.1)
-              soundManager.playStepSound(8);
-            if (currAniTime > 2.292 && currAniTime < 2.4)
-              soundManager.playStepSound(9);
-          }
-        }
-      } */
-      
-      // crouch
-      // const keyOther = _getAnimationKey(true);
-      const keyAnimationAnglesOther = getClosest2AnimationAngles('crouch', angle);
-      const keyAnimationAnglesOtherMirror = _getMirrorAnimationAngles(keyAnimationAnglesOther, 'crouch');
-      const idleAnimationOther = _getIdleAnimation('crouch');
-      
-      const angleToClosestAnimation = Math.abs(angleDifference(angle, keyWalkAnimationAnglesMirror[0].angle));
-      const angleBetweenAnimations = Math.abs(angleDifference(keyWalkAnimationAnglesMirror[0].angle, keyWalkAnimationAnglesMirror[1].angle));
-      const angleFactor = (angleBetweenAnimations - angleToClosestAnimation) / angleBetweenAnimations;
-      const isBackward = _getAngleToBackwardAnimation(keyWalkAnimationAnglesMirror) < Math.PI*0.4;
-      if (isBackward !== this.lastIsBackward) {
-        this.backwardAnimationSpec = {
-          startFactor: this.lastBackwardFactor,
-          endFactor: isBackward ? 1 : 0,
-          startTime: now,
-          endTime: now + 150,
-        };
-        this.lastIsBackward = isBackward;
-      }
-      let mirrorFactor;
-      if (this.backwardAnimationSpec) {
-        const f = (now - this.backwardAnimationSpec.startTime) / (this.backwardAnimationSpec.endTime - this.backwardAnimationSpec.startTime);
-        if (f >= 1) {
-          mirrorFactor = this.backwardAnimationSpec.endFactor;
-          this.backwardAnimationSpec = null;
-        } else {
-          mirrorFactor = this.backwardAnimationSpec.startFactor +
-            Math.pow(
-              f,
-              0.5
-            ) * (this.backwardAnimationSpec.endFactor - this.backwardAnimationSpec.startFactor);
-        }
-      } else {
-        mirrorFactor = isBackward ? 1 : 0;
-      }
-      this.lastBackwardFactor = mirrorFactor;
-
-      const _getHorizontalBlend = (k, lerpFn, isPosition, target) => {
-        _get7wayBlend(
-          keyWalkAnimationAngles,
-          keyWalkAnimationAnglesMirror,
-          keyRunAnimationAngles,
-          keyRunAnimationAnglesMirror,
-          idleAnimation,
-          // mirrorFactor,
-          // angleFactor,
-          // walkRunFactor,
-          // idleWalkFactor,
-          k,
-          lerpFn,
-          isPosition,
-          localQuaternion
-        );
-        _get7wayBlend(
-          keyAnimationAnglesOther,
-          keyAnimationAnglesOtherMirror,
-          keyAnimationAnglesOther,
-          keyAnimationAnglesOtherMirror,
-          idleAnimationOther,
-          // mirrorFactor,
-          // angleFactor,
-          // walkRunFactor,
-          // idleWalkFactor,
-          k,
-          lerpFn,
-          isPosition,
-          localQuaternion2
-        );
-        
-        //_get5wayBlend(keyAnimationAnglesOther, keyAnimationAnglesOtherMirror, idleAnimationOther, mirrorFactor, angleFactor, speedFactor, k, lerpFn, localQuaternion2);
-        
-        lerpFn
-          .call(
-            target.copy(localQuaternion),
-            localQuaternion2,
-            crouchFactor
-          );
-
-      };
-      const _handleDefault = spec => {
-        const {
-          animationTrackName: k,
-          dst,
-          // isTop,
-          lerpFn,
-          isPosition,
-        } = spec;
-        
-        _getHorizontalBlend(k, lerpFn, isPosition, dst);
-      };
-      const _getApplyFn = () => {
-        if (this.jumpState) {
-          return spec => {
-            const {
-              animationTrackName: k,
-              dst,
-              // isTop,
-            } = spec;
-
-            const t2 = this.jumpTime/1000 * 0.6 + 0.7;
-            const src2 = jumpAnimation.interpolants[k];
-            const v2 = src2.evaluate(t2);
-
-            dst.fromArray(v2);
-          };
-        }
-        if (this.sitState) {
-          return spec => {
-            const {
-              animationTrackName: k,
-              dst,
-              // isTop,
-            } = spec;
-            
-            const sitAnimation = sitAnimations[this.sitAnimation || defaultSitAnimation];
-            const src2 = sitAnimation.interpolants[k];
-            const v2 = src2.evaluate(1);
-
-            dst.fromArray(v2);
-          }
-        }
-        if (this.narutoRunState) {
-          return spec => {
-            const {
-              animationTrackName: k,
-              dst,
-              // isTop,
-              isPosition,
-            } = spec;
-            
-            const narutoRunAnimation = narutoRunAnimations[defaultNarutoRunAnimation];
-            const src2 = narutoRunAnimation.interpolants[k];
-            const t2 = (this.narutoRunTime / 1000 * narutoRunTimeFactor) % narutoRunAnimation.duration;
-            const v2 = src2.evaluate(t2);
-
-            dst.fromArray(v2);
-
-            _clearXZ(dst, isPosition);
-          };
-        }
-
-        if (this.danceTime > 0) {
-          return spec => {
-            const {
-              animationTrackName: k,
-              dst,
-              lerpFn,
-              // isTop,
-              isPosition,
-            } = spec;
-
-            _handleDefault(spec);
-
-            const danceAnimation = danceAnimations[this.danceAnimation || defaultDanceAnimation];
-            const src2 = danceAnimation.interpolants[k];
-            const t2 = (timestamp/1000) % danceAnimation.duration;
-            const v2 = src2.evaluate(t2);
-
-            const danceTimeS = this.danceTime/crouchMaxTime;
-            const f = Math.min(Math.max(danceTimeS, 0), 1);
-            lerpFn
-              .call(
-                dst,
-                localQuaternion.fromArray(v2),
-                f
-              );
-
-            _clearXZ(dst, isPosition);
-          };
-        }
-
-        /* if (this.fallLoopState) {
-          return spec => {
-            const {
-              animationTrackName: k,
-              dst,
-              // isTop,
-            } = spec;
-
-            const t2 = (this.fallLoopTime/1000) ;
-            const src2 = fallLoop.interpolants[k];
-            const v2 = src2.evaluate(t2);
-
-            dst.fromArray(v2);
-          };
-        } */
-        if (
-          this.useAnimation ||
-          this.useAnimationCombo.length > 0 ||
-          this.useAnimationEnvelope.length > 0
-        ) {
-          return spec => {
-            const {
-              animationTrackName: k,
-              dst,
-              // isTop,
-              isPosition,
-            } = spec;
-
-            let useAnimation;
-            let t2;
-            const useTimeS = this.useTime/1000;
-            if (this.useAnimation) {
-              const useAnimationName = this.useAnimation;
-              useAnimation = useAnimations[useAnimationName];
-              t2 = Math.min(useTimeS, useAnimation.duration);
-            } else if (this.useAnimationCombo.length > 0) {
-              const useAnimationName = this.useAnimationCombo[this.useAnimationIndex];
-              useAnimation = useAnimations[useAnimationName];
-              t2 = Math.min(useTimeS, useAnimation.duration);
-            } else if (this.useAnimationEnvelope.length > 0) {
-              let totalTime = 0;
-              for (let i = 0; i < this.useAnimationEnvelope.length - 1; i++) {
-                const animationName = this.useAnimationEnvelope[i];
-                const animation = useAnimations[animationName];
-                totalTime += animation.duration;
-              }
-              
-              if (totalTime > 0) {
-                let animationTimeBase = 0;
-                for (let i = 0; i < this.useAnimationEnvelope.length - 1; i++) {
-                  const animationName = this.useAnimationEnvelope[i];
-                  const animation = useAnimations[animationName];
-                  if (useTimeS < (animationTimeBase + animation.duration)) {
-                    useAnimation = animation;
-                    break;
-                  }
-                  animationTimeBase += animation.duration;
-                }
-                if (useAnimation !== undefined) { // first iteration
-                  t2 = Math.min(useTimeS - animationTimeBase, useAnimation.duration);
-                } else { // loop
-                  const secondLastAnimationName = this.useAnimationEnvelope[this.useAnimationEnvelope.length - 2];
-                  useAnimation = useAnimations[secondLastAnimationName];
-                  t2 = (useTimeS - animationTimeBase) % useAnimation.duration;
-                }
-              }
-            }
-
-            _handleDefault(spec);
-            
-            if (useAnimation) {
-              if (!isPosition) {
-                const src2 = useAnimation.interpolants[k];
-                const v2 = src2.evaluate(t2);
-
-                const idleAnimation = _getIdleAnimation('walk');
-                const t3 = 0;
-                const src3 = idleAnimation.interpolants[k];
-                const v3 = src3.evaluate(t3);
-
-                dst
-                  .premultiply(localQuaternion2.fromArray(v3).invert())
-                  .premultiply(localQuaternion2.fromArray(v2));
-              } else {
-                const src2 = useAnimation.interpolants[k];
-                const v2 = src2.evaluate(t2);
-                localVector2.fromArray(v2);
-                _clearXZ(localVector2, isPosition);
-
-                const idleAnimation = _getIdleAnimation('walk');
-                const t3 = 0;
-                const src3 = idleAnimation.interpolants[k];
-                const v3 = src3.evaluate(t3);
-                localVector3.fromArray(v3);
-                
-                dst
-                  .sub(localVector3)
-                  .add(localVector2);
-              }
-            }
-          };
-        } else if (this.hurtAnimation) {
-          return spec => {
-            const {
-              animationTrackName: k,
-              dst,
-              // isTop,
-              isPosition,
-            } = spec;
-            
-            const hurtAnimation = (this.hurtAnimation && hurtAnimations[this.hurtAnimation]);
-            _handleDefault(spec);
-            const hurtTimeS = this.hurtTime/1000;
-            const t2 = Math.min(hurtTimeS, hurtAnimation.duration);
-            // console.log('hurtAnimation', this.hurtAnimation, this.hurtTime, hurtAnimation.duration, hurtTimeS, t2);
-            if (!isPosition) {
-              if (hurtAnimation) {
-                const src2 = hurtAnimation.interpolants[k];
-                const v2 = src2.evaluate(t2);
-
-                const idleAnimation = _getIdleAnimation('walk');
-                const t3 = 0;
-                const src3 = idleAnimation.interpolants[k];
-                const v3 = src3.evaluate(t3);
-                
-                dst
-                  .premultiply(localQuaternion2.fromArray(v3).invert())
-                  .premultiply(localQuaternion2.fromArray(v2));
-              }
-            } else {
-              const src2 = hurtAnimation.interpolants[k];
-              const v2 = src2.evaluate(t2);
-
-              const idleAnimation = _getIdleAnimation('walk');
-              const t3 = 0;
-              const src3 = idleAnimation.interpolants[k];
-              const v3 = src3.evaluate(t3);
-
-              dst
-                .sub(localVector2.fromArray(v3))
-                .add(localVector2.fromArray(v2));
-            }
-          };
-        } else if (this.aimAnimation) {
-          return spec => {
-            const {
-              animationTrackName: k,
-              dst,
-              // isTop,
-              isPosition,
-            } = spec;
-            
-            const aimAnimation = (this.aimAnimation && aimAnimations[this.aimAnimation]);
-            _handleDefault(spec);
-            const t2 = (this.aimTime/aimMaxTime) % aimAnimation.duration;
-            if (!isPosition) {
-              if (aimAnimation) {
-                const src2 = aimAnimation.interpolants[k];
-                const v2 = src2.evaluate(t2);
-
-                const idleAnimation = _getIdleAnimation('walk');
-                const t3 = 0;
-                const src3 = idleAnimation.interpolants[k];
-                const v3 = src3.evaluate(t3);
-                
-                dst
-                  .premultiply(localQuaternion2.fromArray(v3).invert())
-                  .premultiply(localQuaternion2.fromArray(v2));
-              }
-            } else {
-              const src2 = aimAnimation.interpolants[k];
-              const v2 = src2.evaluate(t2);
-
-              const idleAnimation = _getIdleAnimation('walk');
-              const t3 = 0;
-              const src3 = idleAnimation.interpolants[k];
-              const v3 = src3.evaluate(t3);
-
-              dst
-                .sub(localVector2.fromArray(v3))
-                .add(localVector2.fromArray(v2));
-            }
-          };
-        } else if (this.unuseAnimation && this.unuseTime >= 0) {
-          return spec => {
-            const {
-              animationTrackName: k,
-              dst,
-              lerpFn,
-              // isTop,
-              isPosition,
-            } = spec;
-
-            _handleDefault(spec);
-            
-            const unuseTimeS = this.unuseTime/1000;
-            const unuseAnimationName = this.unuseAnimation;
-            const unuseAnimation = useAnimations[unuseAnimationName];
-            const t2 = Math.min(unuseTimeS, unuseAnimation.duration);
-            const f = Math.min(Math.max(unuseTimeS / unuseAnimation.duration, 0), 1);
-            const f2 = Math.pow(1 - f, 2);
-
-            if (!isPosition) {
-              const src2 = unuseAnimation.interpolants[k];
-              const v2 = src2.evaluate(t2);
-
-              const idleAnimation = _getIdleAnimation('walk');
-              const t3 = 0;
-              const src3 = idleAnimation.interpolants[k];
-              const v3 = src3.evaluate(t3);
-
-              localQuaternion.copy(dst)
-                .premultiply(localQuaternion2.fromArray(v3).invert())
-                .premultiply(localQuaternion2.fromArray(v2));
-
-              lerpFn
-                .call(
-                  dst,
-                  localQuaternion,
-                  f2
-                );
-            } else {
-              const src2 = unuseAnimation.interpolants[k];
-              const v2 = src2.evaluate(t2);
-
-              const idleAnimation = _getIdleAnimation('walk');
-              const t3 = 0;
-              const src3 = idleAnimation.interpolants[k];
-              const v3 = src3.evaluate(t3);
-
-              localVector.copy(dst)
-                .sub(localVector2.fromArray(v3))
-                .add(localVector2.fromArray(v2));
-              
-              lerpFn
-                .call(
-                  dst,
-                  localVector,
-                  f2
-                );
-            }
-
-            if (f >= 1) {
-              this.useAnimation = '';
-            }
-          };
-        }
-        return _handleDefault;
-      };
-      const applyFn = _getApplyFn();
-      const _blendFly = spec => {
-        const {
-          animationTrackName: k,
-          dst,
-          // isTop,
-          lerpFn,
-        } = spec;
-        
-        if (this.flyState || (this.flyTime >= 0 && this.flyTime < 1000)) {
-          const t2 = this.flyTime/1000;
-          const f = this.flyState ? Math.min(cubicBezier(t2), 1) : (1 - Math.min(cubicBezier(t2), 1));
-          const src2 = floatAnimation.interpolants[k];
-          const v2 = src2.evaluate(t2 % floatAnimation.duration);
-
-          lerpFn
-            .call(
-              dst,
-              localQuaternion.fromArray(v2),
-              f
-            );
-        }
-      };
-
-      const _blendActivateAction = spec => {
-        const {
-          animationTrackName: k,
-          dst,
-          // isTop,
-          lerpFn,
-        } = spec;
-        
-        if (this.activateTime > 0) {
-          
-            const localPlayer = metaversefile.useLocalPlayer();
-
-            let defaultAnimation = 'grab_forward';
-
-            if (localPlayer.getAction('activate').animationName) {
-              defaultAnimation = localPlayer.getAction('activate').animationName;
-            }
-
-            const activateAnimation = activateAnimations[defaultAnimation].animation;
-            const src2 = activateAnimation.interpolants[k];
-            const t2 = ((this.activateTime / 1000) * activateAnimations[defaultAnimation].speedFactor) % activateAnimation.duration;
-            const v2 = src2.evaluate(t2);
-
-            const f = this.activateTime > 0 ? Math.min(cubicBezier(t2), 1) : (1 - Math.min(cubicBezier(t2), 1));
-
-            lerpFn
-              .call(
-                dst,
-                localQuaternion.fromArray(v2),
-                f
-              );
-        }
-      };
-
-      for (const spec of this.animationMappings) {
-        const {
-          animationTrackName: k,
-          dst,
-          // isTop,
-          isPosition,
-        } = spec;
-        
-        applyFn(spec);
-        _blendFly(spec);
-        _blendActivateAction(spec);
-        
-        // ignore all animation position except y
-        if (isPosition) {
-          if (!this.jumpState) {
-            // animations position is height-relative
-            dst.y *= this.height; // XXX this could be made perfect by measuring from foot to hips instead
-          } else {
-            // force height in the jump case to overide the animation
-            dst.y = this.height * 0.55;
-          }
-        }
-      }
-    };
-    _applyAnimation();
     
     const _overwritePose = poseName => {
       const poseAnimation = animations.index[poseName];
@@ -2898,14 +1425,14 @@ class Avatar {
       // const noiseTime = (now/1000) % noiseAnimation.duration;
       for (const spec of this.animationMappings) {
         const {
-          animationTrackName: k,
+          animationTrackName: atn,
           dst,
           // isTop,
           isPosition,
         } = spec;
 
         if (!isPosition) {
-          const src = poseAnimation.interpolants[k];
+          const src = poseAnimation.interpolants[atn];
           const v = src.evaluate(0);
           dst.fromArray(v);
         }
@@ -2924,139 +1451,60 @@ class Avatar {
         )
       ); */
     };
-    if (this.poseAnimation) {
-      _overwritePose(this.poseAnimation);
-    }
-
-    if (this.getTopEnabled() || this.getHandEnabled(0) || this.getHandEnabled(1)) {
-      this.sdkInputs.hmd.position.copy(this.inputs.hmd.position);
-      this.sdkInputs.hmd.quaternion.copy(this.inputs.hmd.quaternion);
-      this.sdkInputs.leftGamepad.position.copy(this.inputs.leftGamepad.position).add(localVector.copy(this.handOffsetLeft).applyQuaternion(this.inputs.leftGamepad.quaternion));
-      this.sdkInputs.leftGamepad.quaternion.copy(this.inputs.leftGamepad.quaternion);
-      this.sdkInputs.leftGamepad.pointer = this.inputs.leftGamepad.pointer;
-      this.sdkInputs.leftGamepad.grip = this.inputs.leftGamepad.grip;
-      this.sdkInputs.rightGamepad.position.copy(this.inputs.rightGamepad.position).add(localVector.copy(this.handOffsetRight).applyQuaternion(this.inputs.rightGamepad.quaternion));
-      this.sdkInputs.rightGamepad.quaternion.copy(this.inputs.rightGamepad.quaternion);
-      this.sdkInputs.rightGamepad.pointer = this.inputs.rightGamepad.pointer;
-      this.sdkInputs.rightGamepad.grip = this.inputs.rightGamepad.grip;
-
-      const modelScaleFactor = this.sdkInputs.hmd.scaleFactor;
-      if (modelScaleFactor !== this.lastModelScaleFactor) {
-        this.model.scale.set(modelScaleFactor, modelScaleFactor, modelScaleFactor);
-        this.lastModelScaleFactor = modelScaleFactor;
-
-        this.springBoneManager && this.springBoneManager.springBoneGroupList.forEach(springBoneGroup => {
-          springBoneGroup.forEach(springBone => {
-            springBone._worldBoneLength = springBone.bone
-              .localToWorld(localVector.copy(springBone._initialLocalChildPosition))
-              .sub(springBone._worldPosition)
-              .length();
-          });
-        });
-      }
-      
-      if (this.options.fingers) {
-        const _traverse = (o, fn) => {
-          fn(o);
-          for (const child of o.children) {
-            _traverse(child, fn);
-          }
-        };
-        const _processFingerBones = left => {
-          const fingerBones = left ? this.fingerBoneMap.left : this.fingerBoneMap.right;
-          const gamepadInput = left ? this.sdkInputs.leftGamepad : this.sdkInputs.rightGamepad;
-          for (const fingerBone of fingerBones) {
-            // if (fingerBone) {
-              const {bones, finger} = fingerBone;
-              let setter;
-              if (finger === 'thumb') {
-                setter = (q, i) => q.setFromAxisAngle(localVector.set(0, left ? -1 : 1, 0), gamepadInput.grip * Math.PI*(i === 0 ? 0.125 : 0.25));
-              } else if (finger === 'index') {
-                setter = (q, i) => q.setFromAxisAngle(localVector.set(0, 0, left ? 1 : -1), gamepadInput.pointer * Math.PI*0.5);
-              } else {
-                setter = (q, i) => q.setFromAxisAngle(localVector.set(0, 0, left ? 1 : -1), gamepadInput.grip * Math.PI*0.5);
-              }
-              for (let i = 0; i < bones.length; i++) {
-                setter(bones[i].quaternion, i);
-              }
-            // }
-          }
-        };
-        _processFingerBones(true);
-        _processFingerBones(false);
-      }
-    }
-    // if (!this.getBottomEnabled()) {
-      localEuler.setFromQuaternion(this.inputs.hmd.quaternion, 'YXZ');
-      localEuler.x = 0;
-      localEuler.z = 0;
-      localEuler.y += Math.PI;
-      this.modelBoneOutputs.Root.quaternion.setFromEuler(localEuler);
-      
-      this.modelBoneOutputs.Root.position.copy(this.inputs.hmd.position)
-        .sub(localVector.set(0, this.height, 0));
-    // }
-    /* if (!this.getTopEnabled() && this.debugMeshes) {
-      this.modelBoneOutputs.Hips.updateMatrixWorld();
-    } */
-
-    this.shoulderTransforms.Update();
-    this.legsManager.Update();
 
     const _updateEyeTarget = () => {
       const eyePosition = getEyePosition(this.modelBones);
-      const globalQuaternion = localQuaternion2.setFromRotationMatrix(
+      const globalQuaternion = this.localQuaternion2.setFromRotationMatrix(
         this.eyeTargetInverted ?
-          localMatrix.lookAt(
+          this.localMatrix.lookAt(
             this.eyeTarget,
             eyePosition,
             upVector
           )
-        :
-          localMatrix.lookAt(
+          :
+          this.localMatrix.lookAt(
             eyePosition,
             this.eyeTarget,
             upVector
           )
       );
       // this.modelBoneOutputs.Root.updateMatrixWorld();
-      this.modelBoneOutputs.Neck.matrixWorld.decompose(localVector, localQuaternion, localVector2);
+      this.modelBoneOutputs.Neck.matrixWorld.decompose(this.localVector, this.localQuaternion, this.localVector2);
 
-      const needsEyeTarget = this.eyeTargetEnabled && this.modelBones.Root.quaternion.angleTo(globalQuaternion) < Math.PI*0.4;
+      const needsEyeTarget = this.eyeTargetEnabled && this.modelBones.Root.quaternion.angleTo(globalQuaternion) < Math.PI * 0.4;
       if (needsEyeTarget && !this.lastNeedsEyeTarget) {
-        this.startEyeTargetQuaternion.copy(localQuaternion);
+        this.startEyeTargetQuaternion.copy(this.localQuaternion);
         this.lastEyeTargetTime = now;
       } else if (this.lastNeedsEyeTarget && !needsEyeTarget) {
-        this.startEyeTargetQuaternion.copy(localQuaternion);
+        this.startEyeTargetQuaternion.copy(this.localQuaternion);
         this.lastEyeTargetTime = now;
       }
       this.lastNeedsEyeTarget = needsEyeTarget;
 
       const eyeTargetFactor = Math.min(Math.max((now - this.lastEyeTargetTime) / maxEyeTargetTime, 0), 1);
       if (needsEyeTarget) {
-        localQuaternion.copy(this.startEyeTargetQuaternion)
+        this.localQuaternion.copy(this.startEyeTargetQuaternion)
           .slerp(globalQuaternion, cubicBezier(eyeTargetFactor));
-        this.modelBoneOutputs.Neck.matrixWorld.compose(localVector, localQuaternion, localVector2)
+        this.modelBoneOutputs.Neck.matrixWorld.compose(this.localVector, this.localQuaternion, this.localVector2)
         this.modelBoneOutputs.Neck.matrix.copy(this.modelBoneOutputs.Neck.matrixWorld)
-          .premultiply(localMatrix2.copy(this.modelBoneOutputs.Neck.parent.matrixWorld).invert())
-          .decompose(this.modelBoneOutputs.Neck.position, this.modelBoneOutputs.Neck.quaternion, localVector2);
+          .premultiply(this.localMatrix2.copy(this.modelBoneOutputs.Neck.parent.matrixWorld).invert())
+          .decompose(this.modelBoneOutputs.Neck.position, this.modelBoneOutputs.Neck.quaternion, this.localVector2);
       } else {
         if (eyeTargetFactor < 1) {
-          localQuaternion2.copy(this.startEyeTargetQuaternion)
-            .slerp(localQuaternion, cubicBezier(eyeTargetFactor));
-          localMatrix.compose(localVector.set(0, 0, 0), localQuaternion2, localVector2.set(1, 1, 1))
-            .premultiply(localMatrix2.copy(this.modelBoneOutputs.Neck.parent.matrixWorld).invert())
-            .decompose(localVector, localQuaternion, localVector2);
+          this.localQuaternion2.copy(this.startEyeTargetQuaternion)
+            .slerp(this.localQuaternion, cubicBezier(eyeTargetFactor));
+          this.localMatrix.compose(this.localVector.set(0, 0, 0), this.localQuaternion2, this.localVector2.set(1, 1, 1))
+            .premultiply(this.localMatrix2.copy(this.modelBoneOutputs.Neck.parent.matrixWorld).invert())
+            .decompose(this.localVector, this.localQuaternion, this.localVector2);
         }
       }
     };
-    _updateEyeTarget();
 
     const _updateEyeballTarget = () => {
       const leftEye = this.modelBoneOutputs['Eye_L'];
       const rightEye = this.modelBoneOutputs['Eye_R'];
 
-      const lookerEyeballTarget = this.looker.update(now);
+      const lookerEyeballTarget = this.looker.update(now, this);
       const eyeballTarget = this.eyeballTargetEnabled ? this.eyeballTarget : lookerEyeballTarget;
 
       if (eyeballTarget && this.firstPersonCurves) {
@@ -3071,14 +1519,14 @@ class Avatar {
           const headPosition = eyePosition;
 
           // Look at direction in world coordinate
-          const lookAtDir = localVector2.copy(headPosition).sub(position).normalize();
-      
+          const lookAtDir = this.localVector2.copy(headPosition).sub(position).normalize();
+
           // Transform the direction into local coordinate from the first person bone
           lookAtDir.applyQuaternion(
-            localQuaternion.setFromRotationMatrix(head.matrixWorld)
+            this.localQuaternion.setFromRotationMatrix(head.matrixWorld)
               .invert()
           );
-      
+
           // convert the direction into euler
           target.x = Math.atan2(lookAtDir.y, Math.sqrt(lookAtDir.x * lookAtDir.x + lookAtDir.z * lookAtDir.z));
           target.y = Math.atan2(-lookAtDir.x, -lookAtDir.z);
@@ -3088,64 +1536,64 @@ class Avatar {
         function lookAtEuler(euler) {
           const srcX = euler.x;
           const srcY = euler.y;
-      
+
           const rotationFactor = Math.PI;
-          const rotationRange = Math.PI*0.1;
+          const rotationRange = Math.PI * 0.1;
 
           if (leftEye) {
             if (srcX < 0.0) {
-              localEuler2.x = -lookAtVerticalDownCurve.map(-srcX);
+              this.localEulerHmd.x = -lookAtVerticalDownCurve.map(-srcX);
             } else {
-              localEuler2.x = lookAtVerticalUpCurve.map(srcX);
+              this.localEulerHmd.x = lookAtVerticalUpCurve.map(srcX);
             }
-      
-            if (srcY < 0.0) {
-              localEuler2.y = -lookAtHorizontalInnerCurve.map(-srcY);
-            } else {
-              localEuler2.y = lookAtHorizontalOuterCurve.map(srcY);
-            }
-            localEuler2.x *= rotationFactor;
-            localEuler2.y *= rotationFactor;
-            localEuler2.y = Math.min(Math.max(localEuler2.y, -rotationRange), rotationRange);
 
-            localEuler2.z = 0;
-            localEuler2.order = 'YXZ';
-      
-            leftEye.quaternion.setFromEuler(localEuler2);
+            if (srcY < 0.0) {
+              this.localEulerHmd.y = -lookAtHorizontalInnerCurve.map(-srcY);
+            } else {
+              this.localEulerHmd.y = lookAtHorizontalOuterCurve.map(srcY);
+            }
+            this.localEulerHmd.x *= rotationFactor;
+            this.localEulerHmd.y *= rotationFactor;
+            this.localEulerHmd.y = Math.min(Math.max(this.localEulerHmd.y, -rotationRange), rotationRange);
+
+            this.localEulerHmd.z = 0;
+            this.localEulerHmd.order = 'YXZ';
+
+            leftEye.quaternion.setFromEuler(this.localEulerHmd);
             leftEye.updateMatrix();
           }
-      
+
           if (rightEye) {
             if (srcX < 0.0) {
-              localEuler2.x = -lookAtVerticalDownCurve.map(-srcX);
+              this.localEulerHmd.x = -lookAtVerticalDownCurve.map(-srcX);
             } else {
-              localEuler2.x = lookAtVerticalUpCurve.map(srcX);
+              this.localEulerHmd.x = lookAtVerticalUpCurve.map(srcX);
             }
-      
-            if (srcY < 0.0) {
-              localEuler2.y = -lookAtHorizontalOuterCurve.map(-srcY);
-            } else {
-              localEuler2.y = lookAtHorizontalInnerCurve.map(srcY);
-            }
-            localEuler2.x *= rotationFactor;
-            localEuler2.y *= rotationFactor;
-            localEuler2.y = Math.min(Math.max(localEuler2.y, -rotationRange), rotationRange);
 
-            localEuler2.z = 0;
-            localEuler2.order = 'YXZ';
-      
-            rightEye.quaternion.setFromEuler(localEuler2);
+            if (srcY < 0.0) {
+              this.localEulerHmd.y = -lookAtHorizontalOuterCurve.map(-srcY);
+            } else {
+              this.localEulerHmd.y = lookAtHorizontalInnerCurve.map(srcY);
+            }
+            this.localEulerHmd.x *= rotationFactor;
+            this.localEulerHmd.y *= rotationFactor;
+            this.localEulerHmd.y = Math.min(Math.max(this.localEulerHmd.y, -rotationRange), rotationRange);
+
+            this.localEulerHmd.z = 0;
+            this.localEulerHmd.order = 'YXZ';
+
+            rightEye.quaternion.setFromEuler(this.localEulerHmd);
             rightEye.updateMatrix();
           }
         }
         function lookAt(position) {
-          calculateTargetEuler(localEuler, position);
-          lookAtEuler(localEuler);
+          calculateTargetEuler.call(this, this.localEuler, position);
+          lookAtEuler.call(this, this.localEuler);
         }
 
         const head = this.modelBoneOutputs.Head;
         const eyePosition = getEyePosition(this.modelBones);
-        lookAt(eyeballTarget);
+        lookAt.call(this, eyeballTarget);
       } else {
         if (leftEye) {
           leftEye.quaternion.identity();
@@ -3157,25 +1605,6 @@ class Avatar {
         }
       }
     };
-    _updateEyeballTarget();
-
-    this.modelBoneOutputs.Root.updateMatrixWorld();
-    
-    Avatar.applyModelBoneOutputs(
-      this.foundModelBones,
-      this.modelBoneOutputs,
-      // this.getTopEnabled(),
-      this.getBottomEnabled(),
-      this.getHandEnabled(0),
-      this.getHandEnabled(1),
-    );
-    // this.modelBones.Root.updateMatrixWorld();
-
-    this.springBoneTimeStep.update(timeDiff);
-
-    // XXX hook these up
-    this.nodder.update(now);
-    this.emoter.update(now);
 
     const _updateVisemes = () => {
       const volumeValue = this.volume !== -1 ? Math.min(this.volume * 12, 1) : -1;
@@ -3200,28 +1629,28 @@ class Avatar {
             // surprisedIndex,
             // extraIndex,
           ] = visemeMapping;
-          
+
           // reset
           for (let i = 0; i < morphTargetInfluences.length; i++) {
             morphTargetInfluences[i] = 0;
           }
-          
+
           // if (volumeValue !== -1) { // real speech
-            if (aIndex !== -1) {
-              morphTargetInfluences[aIndex] = volumeValue;
-            }
-            if (eIndex !== -1) {
-              morphTargetInfluences[eIndex] = volumeValue * this.vowels[1];
-            }
-            if (iIndex !== -1) {
-              morphTargetInfluences[iIndex] = volumeValue * this.vowels[2];
-            }
-            if (oIndex !== -1) {
-              morphTargetInfluences[oIndex] = volumeValue * this.vowels[3];
-            }
-            if (uIndex !== -1) {
-              morphTargetInfluences[uIndex] = volumeValue * this.vowels[4];
-            }
+          if (aIndex !== -1) {
+            morphTargetInfluences[aIndex] = volumeValue;
+          }
+          if (eIndex !== -1) {
+            morphTargetInfluences[eIndex] = volumeValue * this.vowels[1];
+          }
+          if (iIndex !== -1) {
+            morphTargetInfluences[iIndex] = volumeValue * this.vowels[2];
+          }
+          if (oIndex !== -1) {
+            morphTargetInfluences[oIndex] = volumeValue * this.vowels[3];
+          }
+          if (uIndex !== -1) {
+            morphTargetInfluences[uIndex] = volumeValue * this.vowels[4];
+          }
           /* } else { // fake speech
             this.fakeSpeechSmoothed = this.fakeSpeechSmoothed * 0.99 + 0.01 * this.fakeSpeechValue;
             const now2 = now / 1000 * 2;
@@ -3247,7 +1676,7 @@ class Avatar {
               morphTargetInfluences[uIndex] = this.fakeSpeechSmoothed * uValue/totalValue;
             }
           } */
-            
+
           // emotes
           if (this.emotes.length > 0) {
             for (const emote of this.emotes) {
@@ -3294,7 +1723,6 @@ class Avatar {
         }
       }
     };
-    this.options.visemes && _updateVisemes();
 
     const _updateSubAvatars = () => {
       if (this.spriteMegaAvatarMesh) {
@@ -3304,10 +1732,126 @@ class Avatar {
         });
       }
     };
+
+    //_motionControl is currently unused and may end up being deleted
+    const _motionControl = () => {
+      this.sdkInputs.hmd.position.copy(this.inputs.hmd.position);
+      this.sdkInputs.hmd.quaternion.copy(this.inputs.hmd.quaternion);
+      this.sdkInputs.leftGamepad.position.copy(this.inputs.leftGamepad.position).add(this.localVector.copy(this.handOffsetLeft).applyQuaternion(this.inputs.leftGamepad.quaternion));
+      this.sdkInputs.leftGamepad.quaternion.copy(this.inputs.leftGamepad.quaternion);
+      this.sdkInputs.leftGamepad.pointer = this.inputs.leftGamepad.pointer;
+      this.sdkInputs.leftGamepad.grip = this.inputs.leftGamepad.grip;
+      this.sdkInputs.rightGamepad.position.copy(this.inputs.rightGamepad.position).add(this.localVector.copy(this.handOffsetRight).applyQuaternion(this.inputs.rightGamepad.quaternion));
+      this.sdkInputs.rightGamepad.quaternion.copy(this.inputs.rightGamepad.quaternion);
+      this.sdkInputs.rightGamepad.pointer = this.inputs.rightGamepad.pointer;
+      this.sdkInputs.rightGamepad.grip = this.inputs.rightGamepad.grip;
+
+      const modelScaleFactor = this.sdkInputs.hmd.scaleFactor;
+      if (modelScaleFactor !== this.lastModelScaleFactor) {
+        this.model.scale.set(modelScaleFactor, modelScaleFactor, modelScaleFactor);
+        this.lastModelScaleFactor = modelScaleFactor;
+
+        this.springBoneManager && this.springBoneManager.springBoneGroupList.forEach(springBoneGroup => {
+          springBoneGroup.forEach(springBone => {
+            springBone._worldBoneLength = springBone.bone
+              .localToWorld(this.localVector.copy(springBone._initialLocalChildPosition))
+              .sub(springBone._worldPosition)
+              .length();
+          });
+        });
+      }
+
+      if (this.options.fingers) {
+        const _traverse = (o, fn) => {
+          fn(o);
+          for (const child of o.children) {
+            _traverse(child, fn);
+          }
+        };
+        const _processFingerBones = left => {
+          const fingerBones = left ? this.fingerBoneMap.left : this.fingerBoneMap.right;
+          const gamepadInput = left ? this.sdkInputs.leftGamepad : this.sdkInputs.rightGamepad;
+          for (const fingerBone of fingerBones) {
+            // if (fingerBone) {
+            const { bones, finger } = fingerBone;
+            let setter;
+            if (finger === 'thumb') {
+              setter = (q, i) => q.setFromAxisAngle(this.localVector.set(0, left ? -1 : 1, 0), gamepadInput.grip * Math.PI * (i === 0 ? 0.125 : 0.25));
+            } else if (finger === 'index') {
+              setter = (q, i) => q.setFromAxisAngle(this.localVector.set(0, 0, left ? 1 : -1), gamepadInput.pointer * Math.PI * 0.5);
+            } else {
+              setter = (q, i) => q.setFromAxisAngle(this.localVector.set(0, 0, left ? 1 : -1), gamepadInput.grip * Math.PI * 0.5);
+            }
+            for (let i = 0; i < bones.length; i++) {
+              setter(bones[i].quaternion, i);
+            }
+            // }
+          }
+        };
+        _processFingerBones(true);
+        _processFingerBones(false);
+      }
+    }
+
+    _updateHmdPosition();
+    _applyAnimation(this, now, moveFactors);
+
+    if (this.poseAnimation) {
+      _overwritePose(this.poseAnimation);
+    }
+
+    // //sync animations with motion controls
+    // if (this.getTopEnabled() || this.getHandEnabled(0) || this.getHandEnabled(1)) {
+    //   _motionControl.call(this);
+    // }
+
+    // make sure avatar is facing correct direction
+    this.localEuler.setFromQuaternion(this.inputs.hmd.quaternion, 'YXZ');
+    this.localEuler.x = 0;
+    this.localEuler.z = 0;
+    this.localEuler.y += Math.PI;
+    this.modelBoneOutputs.Root.quaternion.setFromEuler(this.localEuler);
+    //make sure we're using the correct hight if we've scaled anything
+    this.modelBoneOutputs.Root.position.copy(this.inputs.hmd.position)
+      .sub(this.localVector.set(0, this.height, 0));
+
+    this.shoulderTransforms.Update();
+    this.legsManager.Update();
+
+
+    _updateEyeTarget();
+    _updateEyeballTarget();
+
+    this.modelBoneOutputs.Root.updateMatrixWorld();
+    
+    Avatar.applyModelBoneOutputs(
+      this.foundModelBones,
+      this.modelBoneOutputs,
+      // this.getTopEnabled(),
+      this.getBottomEnabled(),
+      this.getHandEnabled(0),
+      this.getHandEnabled(1),
+    );
+    // this.modelBones.Root.updateMatrixWorld();
+
+    if (this.springBoneManager) {
+      this.springBoneTimeStep.update(timeDiff);
+    }
+    /* if (this.springBoneManager && wasDecapitated) {
+      this.decapitate();
+    } */
+
+    // XXX hook these up
+    this.nodder.update(now);
+    this.emoter.update(now);
+
+    this.options.visemes && _updateVisemes();
+
+
     _updateSubAvatars();
 
-    if (metaversefile.isDebugMode() && !this.debugMesh) {
-      this.debugMesh = _makeDebugMesh();
+    if (game.debugMode && !this.debugMesh) {
+      this.debugMesh = _makeDebugMesh(this);
       this.debugMesh.wrapToAvatar(this);
       this.model.add(this.debugMesh);
     }
